server/requirement.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
names

